SSPCPP-689 Bump versions and names in Versions.props
https://issues.shibboleth.net/jira/projects/SSPCPP/issues/SSPCPP-689

Also remove any legacy support for pre OpenSSL1.1.

For now we invent a notional Shibboleth2.7 with OpenSAML and XMLTooling
versions of 2.7.  This means that (my) test machines won't get too
polluted with 3.0 versions while the APIs are unstable.

It will be a pretty quick change to move them up to 3.0