SSPCPP-782 Do not put debug DLLs and EXEs into the installer
https://issues.shibboleth.net/jira/browse/SSPCPP-782

Introduce an environmental variable DEBUG_INSTALLER which if not
set to YES will cause the msi and sms files to be built without
the debug dll and exes.

To speed up the build, undtr this configuration we do not build them
for XMLTooling, OpenSAML and Shibboleth.  The dependency build
(currently) builds both variants.